Introduce Logger
The library does not produce any log output because it is not
aware where to produce output to. It fully depends on the client
application. If a client application wants to make the library's
output a part of its own log it should provide a corresponding
log function and set verbosity level.