libmusic: fix overlapping segments returned by chord_detector
Assuming endIdx is an index of the last sample of chord it should be taken into account
when segments list is built. Otherwise calculation of chord's size gets broken, which in
turn leads to bizarre behavior, e.g:

   $ lmclient -c -r 'A#sus2' 'A#sus2.wav'
   4869577.50 %

because size of segment is startIdx - endIds + 1, which erroneously becomes one sample off
the actual segment size, so total sum of segments's sizes is larger than input size.